requirements.txt
# my deps
google-cloud-firestore
cryptography
